feat(post): add like and bookmark counts in response
Update PostController and PostService to include `likeCount` and `bookmarkCount` in post details and lists. Refactor repositories to return counts as `Int`. Update related DTOs and service logic for enhanced response data.

+++ src/main/kotlin/com/example/powerclean/application/port/outbound/persistence/PostLikeRepository.kt
@@ -16,7 +16,7 @@ interface PostLikeRepository {
         accountId: UUID,
     )
 
-    fun countByPostId(postId: UUID): Long
+    fun countByPostId(postId: UUID): Int
 
     fun findAllByAccountId(accountId: UUID): List<PostLike>
 }
