+++ src/main/kotlin/com/example/powerclean/presentation/dto/CountResDto.kt
@@ -1,5 +1,5 @@
 package com.example.powerclean.presentation.dto
 
 data class CountResDto(
-    val count: Long,
+    val count: Int,
 )
